--- a/radis/los/slabs.py
+++ b/radis/los/slabs.py
@@ -96,8 +96,9 @@ def SerialSlabs(*slabs, **kwargs):
     verbose: bool
         if ``True``, more blabla. Default ``False``
     modify_inputs: False
-        if ``True``, slabs are modified directly when they are resampled. This
-        avoids making a copy so is slightly faster. Default ``False``.
+        if ``True``, slabs wavelengths/wavenumbers are modified directly when
+        they are resampled. This avoids making a copy so it is slightly faster.
+        Default ``False``.
 
         ..note::
             for large number of slabs (in radiative transfer calculations) you
@@ -142,7 +143,6 @@ def SerialSlabs(*slabs, **kwargs):
     modify_inputs = kwargs.pop("modify_inputs", False)  # type: bool
     if len(kwargs) > 0:
         raise ValueError("Unexpected input: {0}".format(list(kwargs.keys())))
-
     if resample_wavespace not in ["never", "intersect", "full"]:
         raise ValueError(
             "resample should be one of: {0}".format(
@@ -215,6 +215,7 @@ def SerialSlabs(*slabs, **kwargs):
                 "transmittance_noslit",
                 wunit=waveunit,
                 Iunit=unitsn["transmittance_noslit"],
+                copy=False,
             )[1]
         try:
             sn.update("radiance_noslit", verbose=verbose)
@@ -222,7 +223,10 @@ def SerialSlabs(*slabs, **kwargs):
             pass
         else:
             In = sn.get(
-                "radiance_noslit", wunit=waveunit, Iunit=unitsn["radiance_noslit"]
+                "radiance_noslit",
+                wunit=waveunit,
+                Iunit=unitsn["radiance_noslit"],
+                copy=False,
             )[1]
         # ... get s quantities
         try:
@@ -234,6 +238,7 @@ def SerialSlabs(*slabs, **kwargs):
                 "transmittance_noslit",
                 wunit=waveunit,
                 Iunit=unitsn["transmittance_noslit"],
+                copy=False,
             )[1]
         try:
             s.update("radiance_noslit", verbose=verbose)
@@ -241,7 +246,10 @@ def SerialSlabs(*slabs, **kwargs):
             pass
         else:
             I = s.get(
-                "radiance_noslit", wunit=waveunit, Iunit=unitsn["radiance_noslit"]
+                "radiance_noslit",
+                wunit=waveunit,
+                Iunit=unitsn["radiance_noslit"],
+                copy=False,
             )[1]
 
         # Solve radiative transfer equation
@@ -275,6 +283,7 @@ def SerialSlabs(*slabs, **kwargs):
             cond_units=cond_units,
             units=unitsn,
             name=name,
+            warnings=False,  # we already know waveranges are properly spaced, etc.
         )
 
 
@@ -372,11 +381,16 @@ def resample_slabs(
         resampled copies of inputs Spectra. All now have the same wavespace
     """
 
-    # Check wavespace is the same
     def same_wavespace(wl):
-        if not all([len(w) == len(wl[0]) for w in wl[1:]]):
+        try:
+            assert all([(w == wl[0]).all() for w in wl[1:]])
+        except AssertionError:
+            # ok if wavespace is the same within some tolerance
+            #   @dev: testing with == first is 10x faster and works in most cases.
+            if all([allclose(w, wl[0]) for w in wl[1:]]):
+                return True
             return False
-        elif not all([allclose(w, wl[0]) for w in wl[1:]]):
+        except:  # ex:  different lengths
             return False
         else:
             return True
@@ -396,7 +410,7 @@ def resample_slabs(
         # defined on the same range)
         slabsk = [s for s in slabs if k in s.get_vars()]  # note that these are
         # references to the actual Spectrum copy
-        wl = [s.get(k, wunit=waveunit)[0] for s in slabsk]
+        wl = [s.get(k, wunit=waveunit, copy=False)[0] for s in slabsk]
         if not same_wavespace(wl):
             # resample slabs if allowed
             if resample_wavespace == "never":
@@ -414,14 +428,13 @@ def resample_slabs(
                 wnew = arange(wmin, wmax + dw, dw)
                 if wnew[-1] > wmax:  # sometimes arange doesnt work as expected
                     wnew = wnew[:-1]
-                # ... copy the array of slabs not to modify the input
-                # ... slabs themselves
                 for s in slabsk:
                     s.resample(
                         wnew,
                         unit=waveunit,
                         if_conflict_drop="convoluted",
                         out_of_bounds=out_of_bounds,
+                        inplace=True,  # we already copied 'if not modify_inputs'
                     )
                 # note: s.resample() fills with 0 when out of bounds
             elif resample_wavespace == "intersect":
@@ -434,14 +447,13 @@ def resample_slabs(
                     wnew = wnew[:-1]
                 if len(wnew) == 0:
                     raise ValueError("Intersect range is empty")
-                # ... copy the array of slabs not to modify the input
-                # ... slabs themselves
                 for s in slabsk:
                     s.resample(
                         wnew,
                         unit=waveunit,
                         if_conflict_drop="convoluted",
                         out_of_bounds=out_of_bounds,
+                        inplace=True,  # we already copied 'if not modify_inputs'
                     )
                 # note: s.resample() fills with 0 when out of bounds
         # Now all our slabs have the same wavespace