@erwanp 2021-01-25
@Anjum48  this improvement may help in your case (with ~ 90 layers I'd expect only about 2-3s were spent on useless recursive checks. It would only get bad (tens of seconds) beyond 200 layers. 
 
Also @ulyssed you'll definitly need this if you end up using SerialSlabs for radiative transfer with your 800+ slabs :)  

Will be added in 0.9.28, get the developer version if you need it now !